Backend/requirements.txt
Django==4.2.0
djangorestframework
django-cors-headers
gunicorn
python-dotenv
psycopg2-binary
dj-database-url
django-cors-headers